==> notebooks/1 - Data Extraction - Covid19 Brazil IO Data.ipynb <==
# Extraindo dados da Covid19 via API
### Fonte de Dados: https://brasil.io/dataset/covid19/
### Augusto SPINELLI

### Licença
Os dados dados convertidos estão sob a licença Creative Commons Attribution ShareAlike. Caso utilize os dados, cite a fonte original e quem tratou os dados, como: Fonte: Secretarias de Saúde das Unidades Federativas, dados tratados por Álvaro Justen e colaboradores/Brasil.IO. Caso compartilhe os dados, utilize a mesma licença.

***

## Extração de Dados

#importando lib para realizar consultas na API
import requests

#URL da API
url_estados = 'https://brasil.io/api/dataset/covid19/caso/data?is_last=True&place_type=state'
url_cidades = 'https://brasil.io/api/dataset/covid19/caso/data/?is_last=True&place_type=city&page_size=10000'
url_estados_historico = 'https://brasil.io/api/dataset/covid19/caso/data/?place_type=state&page_size=10000'

#Consulta
result_estados = requests.get(url_estados)
result_cidades = requests.get(url_cidades)
result_estados_hist = requests.get(url_estados_historico)

#função para salvar arquivos no caminho
def saveFiles(file_path,result):
    with open(file_path,'wb') as handle:
        for block in result.iter_content(1024):
            handle.write(block)

#importando lib para acessar OS
import os

#Caminho dos arquivos RAW
raw_data_path = os.path.join(os.path.pardir, 'data', 'raw')
estados_data_path = os.path.join(raw_data_path, 'estados.json')
cidades_data_path = os.path.join(raw_data_path, 'cidades.json')
estados_historico_data_path = os.path.join(raw_data_path, 'estados_historico.json')

#Salvando arquivos no disco
saveFiles(estados_data_path,result_estados)
saveFiles(cidades_data_path,result_cidades)
saveFiles(estados_historico_data_path,result_estados_hist)
saveFiles(ceps_data_path,result_ceps)

#Criando Pandas DataFrame com os Dados

#importando dependencias
import pandas as pd
import numpy as np

#criando dataframes brutos/raw
cidadesRaw_df = pd.read_json(cidades_data_path)
estadosRaw_df = pd.read_json(estados_data_path)
estadosHRaw_df = pd.read_json(estados_historico_data_path)

#transformando dados para dataframes finais com orientação correta
cidades = cidadesRaw_df['results']
cidades_df = pd.read_json(cidades.to_json(),orient='index')

estados = estadosRaw_df['results']
estados_df = pd.read_json(estados.to_json(),orient='index')

estadosH = estadosHRaw_df['results']
df = pd.read_json(estadosH.to_json(),orient='index')

#verificando os dataframes
cidades_df.info()

estados_df.info()
df.info()

### End of Notebook
