==> notebooks/4 - Data Modeling - Building a Predictive Model.ipynb <==
# Extraindo dados da Covid19 via API
### Fonte de Dados: https://brasil.io/dataset/covid19/
### Augusto SPINELLI

### Licença
Os dados dados convertidos estão sob a licença Creative Commons Attribution ShareAlike. Caso utilize os dados, cite a fonte original e quem tratou os dados, como: Fonte: Secretarias de Saúde das Unidades Federativas, dados tratados por Álvaro Justen e colaboradores/Brasil.IO. Caso compartilhe os dados, utilize a mesma licença.

#### Para criação do nosso modelo preditivo que tentará prever se um paciente covid irá a óbito ou não, utilizaremos o dataset de casos individualizados anonimos baseado no cep do paciente

## Data Munging

É o processo de mapeamento e transformação de dados para análise. No nosso caso, temos acesso somente a um dataset que informa a evolução do paciente com base no endereço / CEP de residencia, e tentaremos preparar um modelo que irá prever a evolucão de pacientes ativos de acordo com as áreas de planejamento Cidade do Rio de Janeiro (APS)


#importando dependencias
import pandas as pd
import numpy as np
import os

#Caminho dos arquivos RAW
raw_data_path = os.path.join(os.path.pardir, 'data', 'raw')
data_path = os.path.join(raw_data_path, 'ceps2.csv')

#criando dataframes brutos/raw
raw_df = pd.read_csv(data_path)

#verificando informações do dataframe
raw_df.info()

#renomeando colunas com encoding errado
df = raw_df.rename(columns={"evolu��o": "evolucao", "dt_�bito": "dt_obito"})

#mapeando data de atualizacao de forma correta
def GetData(dataFull):
    d = dataFull.split(';')[0]
    return d

df['dt_registro'] = df['Data_atualiza��o;OBJECTID'].map(lambda x: GetData(x))

#removendo coluna inútil
del df['Data_atualiza��o;OBJECTID']
df.info()

recuperados = df.loc[df.evolucao == 'recuperado',:] 
print('Nuúmero de Recuperados: {0}'.format(len(recuperados)))

ativos = df.loc[df.evolucao == 'ativo',:] 
print('Nuúmero de Ativos: {0}'.format(len(ativos)))

obitos = df.loc[df.evolucao == '�bito',:] 
print('Nuúmero de Óbitos: {0}'.format(len(obitos)))

df.ap_residencia_estadia.unique()

df = pd.get_dummies(df,columns=['ap_residencia_estadia'])

df.info()

df.drop(['dt_obito','dt_notific','dt_inicio_sintomas','bairro_resid__estadia','CEP','dt_registro'], axis=1,inplace=True)

def preencheSobreviveu(evolucao):
    ret = -1
    if evolucao == 'recuperado':
        ret = 1
    elif evolucao == 'ativo':
        ret = -1 #não sabemos - test data
    else:
        ret = 0 #óbito
    return ret
    
df['sobreviveu'] = df['evolucao'].map(lambda x: preencheSobreviveu(x))

df.drop(['evolucao'], axis=1,inplace=True)
df.info()

#reorder columns
columns = [column for column in df.columns if column != 'sobreviveu']
columns = ['sobreviveu'] + columns
df = df[columns]

df.info()

## Salvando dados em arquivo

Nosso pobre modelo por hora considera como critério apenas a AP de residencia do paciente para definir se ele sobrevive ou não. Devemos procurar um outro dataset com dados anonimos e individualizados mais rica, como por exemplo sexo, renda, etc para que esses critérios também sejam considerados pelo nosso modelo na hora de prever se o paciente sobreviverá ou não.

Por hora, com o que temos em mãos, vamos salvar os datasets de test e training 

processed_data_path = os.path.join(os.path.pardir,'data','processed')
write_train_path = os.path.join(processed_data_path,'train.csv')
write_test_path = os.path.join(processed_data_path,'test.csv')

df.loc[df.sobreviveu != -1].to_csv(write_train_path)
columns = [column for column in df.columns if column != 'sobreviveu']
df.loc[df.sobreviveu == -1, columns].to_csv(write_test_path)

### End of Notebook
