==> notebooks/5 - Data Modeling - Working with a predictive model.ipynb <==
# Extraindo dados da Covid19 via API
### Fonte de Dados: https://brasil.io/dataset/covid19/
### Augusto SPINELLI

### Licença
Os dados dados convertidos estão sob a licença Creative Commons Attribution ShareAlike. Caso utilize os dados, cite a fonte original e quem tratou os dados, como: Fonte: Secretarias de Saúde das Unidades Federativas, dados tratados por Álvaro Justen e colaboradores/Brasil.IO. Caso compartilhe os dados, utilize a mesma licença.

## Importando os dados

import pandas as pd
import os
import numpy as np


#set the path of the processed data
processed_data_path = os.path.join(os.path.pardir,'data','processed')
train_file_path = os.path.join(processed_data_path,'train.csv')
test_file_path = os.path.join(processed_data_path,'test.csv')

#create train and test dataframes using pandas
train_df = pd.read_csv(train_file_path,index_col=[0])
test_df = pd.read_csv(test_file_path,index_col=[0])

train_df.info()
test_df.info()

## Data Preparation

Vamos agora pegar nossos dados de training e quebrá-lo de forma a ter 80% en uma matrix e 20% em outro.

No final, vamos comparar a taxa de sobrevivência em ambos splits para ver se são similares

X = train_df.loc[:,'ap_residencia_estadia_1.0':].to_numpy().astype('float') #input matrix
y = train_df['sobreviveu'].ravel() #output array

print (X.shape, y.shape)

#ML train test split 80/20 (80% will be used to train, 20% to test)
from sklearn.model_selection import train_test_split
X_train, X_test, y_train, y_test = train_test_split(X,y,test_size=0.2,random_state=0)
print (X_train.shape, y_train.shape)
print (X_test.shape, y_test.shape)

#average survival in train and test, without ML 
print('mean survival in train: {0:.3f}'.format(np.mean(y_train)))
print('mean survival in test: {0:.3f}'.format(np.mean(y_test)))

## Modelo Base (Baseline Model)

Este modelo é o nosso modelo base. É o mais burro possível e sempre considera que o paciente sobreviverá

import sklearn
#import function
from sklearn.dummy import DummyClassifier

#create a model
model_dummy = DummyClassifier(strategy='most_frequent',random_state=0)

#train a model
model_dummy.fit(X_train,y_train) #input,output params are necessary for fit training function

#calculate model score (baseline considers most frequent data, a.k.a 1 ==  survived covid)
print('score for baseline model : {0:.2f}'.format(model_dummy.score(X_test,y_test)))

#So, without any ML algorithm, if we simply predict that a pacient with covid survives, we will be right in 90% of the cases
#let's try to improve that with ML! 
#First, lets import some performance metrics
from sklearn.metrics import accuracy_score, confusion_matrix, precision_score, recall_score

#accuracy score
print('accuracy for baseline model : {0:.2f}'.format(accuracy_score(y_test,model_dummy.predict(X_test))))
#confusion matrix
print('confusion matrix for baseline model : \n {0}'.format(confusion_matrix(y_test,model_dummy.predict(X_test))))
#precision and recall scores
print('precision for baseline model : {0:.2f}'.format(precision_score(y_test,model_dummy.predict(X_test))))
print('recall for baseline model : {0:.2f}'.format(recall_score(y_test,model_dummy.predict(X_test))))

## Modelo de regressão lógica

#import logistic regression function from sklearn
from sklearn.linear_model import LogisticRegression

#create our model 
model_lr_1 = LogisticRegression(random_state=0)

#train our model
model_lr_1.fit(X_train,y_train)

#evaluate model
print('score for logistic regression - version 1 : {0:.2f}'.format(model_lr_1.score(X_test,y_test)))

#performance metrics

#accuracy score
print('accuracy for logistic regression v1 model : {0:.2f}'.format(accuracy_score(y_test,model_lr_1.predict(X_test))))
#confusion matrix
print('confusion matrix for logistic regression v1 model : \n {0}'.format(confusion_matrix(y_test,model_lr_1.predict(X_test))))
#precision 
print('precision for logistic regression v1 model : {0:.2f}'.format(precision_score(y_test,model_lr_1.predict(X_test))))
#recall
print('recall for logistic regression v1 model : {0:.2f}'.format(recall_score(y_test,model_lr_1.predict(X_test))))

## Otimização de hyperparametros

# base model
model_lr = LogisticRegression(random_state=0)

from sklearn.model_selection import GridSearchCV

parameters = {'C':[1.0,10.0,50.0,100.0,1000.0],'penalty':['l1','l2']}
clf = GridSearchCV(model_lr,param_grid=parameters,cv=3)

clf.fit(X_train,y_train)

clf.best_params_

print('best score: {0:.2f}'.format(clf.best_score_))

#evaluate model
print('score for logistic regression v2 : {0:.2f}'.format(clf.score(X_test,y_test)))

## End of Notebook

#### Não estamos melhorando nosso modelo base. Como dispomos de dados pobres nesse exemplo e o modelo base já tem fidelidade de 90% (a maioria dos pacientes de covid sobrevivem), um modelo que usa regressão lógica não é suficiente para prever com uma fidelidade ainda maior (dados os poucos parâmetros disponíveis) se o paciente irá sobreviver ou não

#### Neste caso, voltamos a prancheta. Precisamos ter ainda mais dados disponíveis como sexo, idade, renda, etc (e não somente ap de residencia) para tentar montar um modelo mais fiel
