==> notebooks/6 - Data Visualization - Map Viz.ipynb <==
# Extraindo dados da Covid19 via API
### Fonte de Dados: https://brasil.io/dataset/covid19/
### Augusto SPINELLI

### Licença
Os dados dados convertidos estão sob a licença Creative Commons Attribution ShareAlike. Caso utilize os dados, cite a fonte original e quem tratou os dados, como: Fonte: Secretarias de Saúde das Unidades Federativas, dados tratados por Álvaro Justen e colaboradores/Brasil.IO. Caso compartilhe os dados, utilize a mesma licença.

#importando dependencias
import pandas as pd
import numpy as np
import os

#Caminho dos arquivos RAW
raw_data_path = os.path.join(os.path.pardir, 'data', 'raw')
data_path = os.path.join(raw_data_path, 'cepsDataset.csv')

#criando dataframes brutos/raw
df = pd.read_csv(data_path, delimiter = ';')

#mapeando data de atualizacao de forma correta
def GetData(dataFull):
    d = dataFull.split(';')[0]
    return d

df['dt_registro'] = df['Data_atualizacao;OBJECTID'].map(lambda x: GetData(x))

#removendo coluna inútil
del df['Data_atualizacao;OBJECTID']

#vendo info do dataframe
df.info()

## Uso do Mapa com Leaflet

O Leaflet é uma biblioteca JavaScript de código aberto amplamente usada para criar aplicativos de mapeamento da web

#testando ipyleaflet
from ipyleaflet import Map, Marker

center = (-22.9980967, -43.2654989)

m = Map(center=center, zoom=13)

m.add_layer(Marker(location=center))

display(m)


## Usando uma API de CEPs Brasileiros para buscar mais dados

### CEP Aberto
É um projeto que visa prover acesso gratuito e construir de maneira colaborativa uma base de dados com os Codigos de Enderecamento Postal (CEP) geolocalizados de todo o Brasil. Colabore melhorando a informação de CEPs que você conhece. Registre-se gratuitamente para ter acesso a API e poder baixar a base de dados. O CEP Aberto possui informação de 1137131 CEPs distribuídos em 10659 cidades e municípios

#usando a api do cep aberto para pesquisar ceps
import requests

token = "##SECRET##" #pegue seu token pessoal após cadastro no CEP Aberto
headers = {'Authorization': 'Token token=%s' % token}


def search_by_cep(cep):
    url = "https://www.cepaberto.com/api/v3/cep?cep=" + cep
    response = requests.get(url, headers=headers)
    return response.json()

def draw_map(latitude, longitude):
    pos = (latitude, longitude)
    m = Map(center=pos, zoom=13)
    m.add_layer(Marker(location=pos))
    display(m)
    

    

#testando acesso a API
search_by_cep("22041001")


## Covid around me

Um site para mostrar os casos de covid ao redor do usuário

import unicodedata

def strip_accents(text):

    try:
        text = unicode(text, 'utf-8')
    except NameError: # unicode is a default on python 3 
        pass

    text = unicodedata.normalize('NFD', text)\
           .encode('ascii', 'ignore')\
           .decode("utf-8")

    return str(text)

digiteCep = 20540195
buscaCep = search_by_cep(str(digiteCep))

bairroCep = strip_accents(buscaCep['bairro'])
print(bairroCep)


casos_sua_rua = df.loc[df.CEP == digiteCep,:] 
print('Casos na sua Rua: {0}'.format(len(casos_sua_rua)))

casos_seu_bairro = df.loc[df.bairro_resid__estadia == bairroCep.upper(),:] 
print('Casos no seu Bairro: {0}'.format(len(casos_seu_bairro)))


draw_map(buscaCep['latitude'],buscaCep['longitude'])

## End of Notebook
