==> notebooks/7 - Map Viz - Covid Around Me.ipynb <==
import os
import pandas
import requests

#Caminho dos arquivos RAW
raw_data_path = os.path.join(os.path.pardir, 'data', 'raw')
filename = os.path.join(raw_data_path, 'dados.csv')

url_csv = "https://opendata.arcgis.com/datasets/b54234c151aa4d01b488dc12aafd5574_0.csv?outSR=4326"

#save the file
with open(filename, 'wb') as f, \
    requests.get(url_csv, stream=True) as r:
    for line in r.iter_lines():
        f.write(line)
#replace colon for new line in file
with open(filename, 'r') as file:
    filedata = file.read()
    filedata = filedata.replace(',', '\n')
with open(filename, 'w') as file:
    file.write(filedata)

columns = ["dt_notific", "dt_inicio_sintomas", "bairro_resid", "ap_resid","evolucao", "dt_obito", "CEP", "dt_atualizacao","manter"]

df = pandas.read_csv(filename, skiprows=[0,1], delimiter=';',names = columns, error_bad_lines=False)

df.info()


df.head()

#mapeando zonas de acordo com ap
def GetZona(ap):
    ret = ""
    if ap == 1.0:
        ret = "Centro"
    elif ap == 2.2:
        ret = "Zona Norte" 
    elif ap == 3.2:
        ret = "Zona Norte" 
    elif ap == 3.3:
        ret = "Zona Norte" 
    elif ap == 3.1:
        ret = "Zona Norte" 
    elif ap == 2.1:
        ret = "Zona Sul"
    else:
        ret = "Zona Oeste" #ap >= 4
    return ret

df['zona'] = df['ap_resid'].map(lambda x: GetZona(x))

#selected zones
centro = len(df[df.zona == 'Centro'])
norte = len(df[df.zona == 'Zona Norte'])
sul = len(df[df.zona == 'Zona Sul'])
oeste = len(df[df.zona == 'Zona Oeste'])

#selected neighborhoods

ipanema = len(df[df.bairro_resid == 'IPANEMA'])
leblon = len(df[df.bairro_resid == 'LEBLON'])
botafogo = len(df[df.bairro_resid == 'BOTAFOGO'])
copa = len(df[df.bairro_resid == 'COPACABANA'])

tijuca = len(df[df.bairro_resid == 'TIJUCA'])
vila = len(df[df.bairro_resid == 'VILA ISABEL'])
                
campo_grande = len(df[df.bairro_resid == 'CAMPO GRANDE'])
bangu = len(df[df.bairro_resid == 'BANGU'])
santa_cruz = len(df[df.bairro_resid == 'SANTA CRUZ'])
barra = len(df[df.bairro_resid == 'BARRA DA TIJUCA'])



import plotly.graph_objects as go


#plotting the chart
fig =go.Figure(go.Sunburst(
    labels = ['Zona Oeste','CAMPO GRANDE','BANGU','SANTA CRUZ','BARRA DA TIJUCA','Zona Sul','IPANEMA','LEBLON', 'BOTAFOGO','COPACABANA', 'Zona Norte', 'TIJUCA', 'VILA ISABEL', 'Centro'],
    parents = ['', 'Zona Oeste','Zona Oeste','Zona Oeste','Zona Oeste','', 'Zona Sul', 'Zona Sul','Zona Sul','Zona Sul','','Zona Norte','Zona Norte',''],
    values=[oeste,campo_grande,bangu,santa_cruz,barra,sul,ipanema,leblon,botafogo,copa,norte,tijuca,vila,centro],
    textinfo='label+value'
))

fig.update_layout( title={
        'text': "Distribuição casos Covid19 por Zonas e bairros selecionados",
        'y':0.9,
        'x':0.5,
        'xanchor': 'center',
        'yanchor': 'top'} #set the title
        )

#to show the figure
fig.show()
